upgrade to 2025.2.2
fairly quick and dirty, tested with SysON.
The "projectID == editingdomainID" assumption no longer holds hence I
had to tweak and adapt. I'm not 100% sure it's done well but it seems to
work.
Had to change the base URI for API to /api/rest/ in order to bypass some
settings made elsewhere (most likely in Sirius Web.